requirements.txt
langchain
langchain-google-genai
streamlit
python-dotenv
PyPDF2
python-docx
pandas